requirements.txt
altgraph==0.17.4
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
db-dtypes==1.4.3
dotenv==0.9.9
fonttools==4.58.5
google-api-core==2.25.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-cloud-bigquery==3.35.0
google-cloud-bigquery-storage==2.32.0
google-cloud-core==2.4.3
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.73.1
idna==3.10
kiwisolver==1.4.8
macholib==1.16.3
matplotlib==3.10.3
numpy==2.3.1
oauthlib==3.3.1
packaging==25.0
pandas==2.3.1
pandas-gbq==0.29.2
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.31.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydata-google-auth==1.9.1
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.8
pyparsing==3.2.3
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
reportlab==4.4.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
scipy==1.16.1
setuptools==80.9.0
shiboken6==6.9.1
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
